backend/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
python-dotenv
qrcode
gunicorn
pyodbc
SQLAlchemy
python-dotenv